dashboard/requirements.txt
pytest
pylint
streamlit
python-dotenv
boto3
psycopg2-binary